+++ CHANGELOG.md
@@ -129,6 +129,7 @@
 
 ### 1.2.0
 
+- Reimplemented Curios support for the Keep Inventory gene (#20)
 - Updated Patchouli support to use item components rather than NBT
 - Updated the Patchouli book to reflect changes since 1.19
 - The Patchouli book is now in the creative tab